[TensorExpr] Make 'Tensor' a value type.
This is another commit in transition from KernelArena memory management.
Tensor is essentially just a pair of <BufPtr, StmtPtr> and we don't need
to dynamically allocate it at all - it's cheap to pass it by value, and
that's what we're switching to in this commit.

After this change nothing uses KernelScope/KernelArena and they can be
safely removed.

ghstack-source-id: 9206ff0ce169cd29cf9c50f6007ac71c413ed4fa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/63586